services/ingestion/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pdfplumber==0.11.0
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.27.0
